Switch gunicorn comms to a unix socket.
Polling for a unix socket to be opened by gunicorn has proven to be
both fully robust via extended local testing and helps the test run a
few seconds faster.

Also switch gevent monkey-patching to being handled by gunicorn, which
appears to be how this is actually done in production setups. With lazy
imports reverted, the test fails as expected; so the monkey patching
by gunicorn is confirmed to be happening.