Lock sdists in more cases for foreign platforms. (#2508)
Previously, when locking for a foreign platform and only an sdist was
available, the lock would fail if the sdist's build backend did not
support the `prepare_metadata_for_build_wheel` PEP-517 hook and this
resulted in building a wheel that was platform specific and incompatible
with the foreign platform. Now, we grab the METADATA from the
incompatible wheel under the same motivation we do so for
`--style universal` locks: in general, the only sane thing to do is 
assume metadata is consistent across all the distributions for a given 
project release (even though we know this is not always true in the
wild; e.g.: PyTorch).