Support `--pre-resolved-dists` resolver.
When building a PEX or creating a venv via `pex3 venv create` you can
now tell Pex to use a set of pre-resolved distributions. This is similar
to using `--no-pypi --find-links ...` except that:
1. Its roughly 3x faster since Pip is not asked to do any resolution.
2. It requires all the distributions specified form an already complete
   resolve.

One way to obtain distributions that meet criteria 2 is to use `pip
download -d ...` or `pip wheel -w ...` to pre-resolve the distributions
you need.

Closes #1907