Re-work testing pyenv management. (#2535)
The new setup allows for safe upgrades by just bumping pyenv Python
versions. Also re-work atomic_directory to support this cleanly, allow
obtaining an independent lock on an atomic directory after it has been
established. This can be useful when reads of the directory contents all
happen in the same code path as the writes.