Add support for sharding tests.
Create a pytest plugin that supports recording test timings and then
sharding tests based on timings. The initial timing baseline is included
and the plugin is made to handle the case of out of date timings by
assuming an average time for any test not found in the timings database.

N.B.: This will be used in a forthcoming change to shard CI integration
tests and bring CI times down by ~30%.