Add the ability to specify the `--pip-log` path. (#2536)
Although `--pip-log` is just an alias for the pre-existing
`--preserve-pip-download-log` option, the option gains the ability to
accept an optional log path value. In addition to this pro-active means
of starting a debuggable Pex session with Pip, the log is also made more
useful in the face of a multi-target resolve by serializing the log on
targets and prefixing log lines per target. The log now also includes
the compatibility tags calculated via `pip -v debug ...` for any 
abbreviated `--platform`s used in the resolve.