Fix intermittent PEP-517 failures. (#2540)
I've not had a report in the wild, but CI intermittently fails in
PEP-517 tests, not finding the json communication file on the read end.
This exposes a race where the PEP-517 process completes before the
temporary file context manager is exited and the communication file,
with the results therein, is deleted before it can be read. Switch from
a temporary file that deletes on context exit (it's amazing this worked
as reliably as it did - it was a bug from day 1!) to one that deletes
only upon the Pex process exit.