Fix `--pip-log` re-use.
Previously, when only 1 resolve target was specified, a `--pip-log`
would be re-used which would append the current Pip log outout with the
prior output. For a normal PEX build, this was potentially confusing,
but for a `pex3 lock` command it could lead to errors since the log is
used to generate the lock.

Fixes #2568