Fix `--venv` cached Python interpreter info. (#2579)
Previously, the work dir path of the `atomic_directory` used to create
the venv would leak into various cached paths in the `PythonInterpreter`
`INTERP-INFO` file. Now, these paths are corrected at creation time.

This is work towards `pex3 cache prune --last-access` which will need to
iterate cached interpreters to find any associated with venvs such that
the interpreter can be pruned when the venv is pruned.