Accept more foreign `--platform` "YOLO-mode" wheels. (#2607)
Previously, when speculatively building a wheel from an sdist for a
foreign platform target, the wheel tags needed to match the foreign
platform target's tags exactly in all cases. For `--complete-platform`
this makes sense, we have complete information about the foreign
platform target's tags, but for an abbreviated `--platform` it does not
since we have abbreviated information that is not enough to know a tag
mismatch definitively signals the wheel will never work on the foreign
platform target. Now only those definitive cases (e.g.: the speculative
wheel is Linux but the foreign abbreviated platform is macOS) are
rejected. This will let some eventually failing wheels though, but the
warning added in #2533 already covers this risk.