Switch vendoring from RedBaron to libcst. (#2654)
RedBaron stopped being maintained in 2019. Although this change was
motivated by wanting to upgrade vendored packaging to 24.2, which
RedBaron could not handle due to new Python syntax introduced since
RedBaron went dormant, I'm proceeding with the change on its own. It
turns out packaging 24.2 is not a currently viable upgrade for other
reasons.