Fix support for CPython development releases. (#2655)
Fix a long-standing bug handling development versions of CPython (any
non-tagged release of the interpreter). These interpreters report a full
version of `X.Y.Z+` and the trailing `+` leads to a non PEP-440
compliant version number. This, in turn, causes issues with the
`packaging` library leading to failures to evaluate markers for these
interpreters which surface as inscrutable Pex errors.

Apply the same workaround as used in packaging >= 24.1. See:
+ https://github.com/pypa/packaging/pull/802
+ https://github.com/pypa/packaging/pull/825