Let PEX scies install PEXes in the PEX_ROOT.
Previously, the PEX_ROOT was synthesized to be inside the nce SCIE_BASE
which led to very long PEX_ROOT cache paths. This foiled PEX scies from
working on Windows in many cases. The Windows case aside, having PEXes
in PEX scies behave like normal PEXes is a goal which this supports.
Although a PEX scie user should be none-the-wiser, the creator of a PEX
scie gets the runtime PEX behavior and debuggability they expect of a
non-scie PEX.