Add `--scie-assets-base-url` & honor `--{proxy,cert}`.
Introduce `--scie-assets-base-url` which can be used to point to a
science asset mirror previously populated with `ptex`, `scie-jump` and
interpreter providers via `science download ...`.

Also fix PEX scie creation to use either of `--proxy` or `--cert` if set
when building scies. Previously, these options were only honored when
downloading the `science` binary itself but not when running it
subsequently to build scies.

Fixes #2810