Support dependency replacement with `--override`. (#2894)
Instead of just accepting PEP-508 requirement strings for `--override`
values, Pex now also accepts values of the form
`<project name>=<requirement>` to override any transitive requirement
on one project with a requirement on a completely different project.

This can be useful when you need to supply a patch to an existing
published project and would prefer to depend on wheels you pre-build
instead of using a VCS source dependency `--override`, which can be slow
to build.

Addresses #2891 and completes the suggested use cases in #2425 for which
`--override` was originally introduced.