Support scopes for `--index` and `--find-links`. (#2903)
This allows scoping package repositories to only certain projects or
certain target environments or an even narrower combination of the two.

In the context of `--style universal` lock creation, this admits
automatically splitting the lock into two or more locked resolves
using different universes of `--indexes` and `--find-links`. This
new splitting support is also extended to any top-level requirements
that are split via markers as well.

This should help along these issues:
+ https://github.com/pantsbuild/pants/issues/18293
+ https://github.com/pantsbuild/pants/issues/18965
+ https://github.com/pantsbuild/pants/issues/20808

And could be used as the basis to help with an issue like:
https://github.com/pantsbuild/pants/issues/22581