Add `--derive-sources-from-requirements-files`. (#2909)
This allows scoping requirement sources via the structure of
requirements files. If any requirements files are specified that contain
`-f` / `--find-links`, `-i` / `--index-url`, or `--extra-index-url`
options, `--derive-sources-from-requirements-files` will automatically
map these repos as the `--source` for the requirements (if any) declared
in the same requirements file.

This may help with issues like:
  https://github.com/pantsbuild/pants/issues/22581