Implement `.whl` packing from chroots and venvs.
Any wheel originating from Pex will be packed to a bytewise identical
wheel. If the wheel originates from either `--pre-resolved-dists` or a
`--venv-repository` that was not created by Pex, the wheel contents will
be bytewise identical, but the wheel itself may have different file
modes and ordering of entries.

Fixes #2391
Fixes #2155
Closes #2299