Fix URL requirement `--lock` & `--pylock` subsetting. (#3034)
Previously, just the project name was matched when subsetting but now
the normalized URL is matched. The previous behavior could lead to
subsets succeeding that should have otherwise failed. The new behavior
can lead to a subset failing when URLs differ, but both URLs point to
the same content. Although this too is a bug, it should be a much
narrower use case in the wild; so this should be an improvement.

Fixes #3032